패키지 구조 변경 com.news->com.news.{프로젝트명}
