+++ discovery-service/src/main/java/com/news/discovery/EurekaApplication.java
@@ -1,4 +1,4 @@
-package com.news;
+package com.news.discovery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
