+++ discovery-service/src/main/java/com/news/discovery/WebSecurityConfig.java
@@ -1,3 +1,4 @@
+package com.news.discovery;
 //package com.news;
 //
 //import org.springframework.security.config.annotation.web.builders.HttpSecurity;
