+++ news-service/.gitignore
@@ -9,3 +9,4 @@
 /.settings/
 /mvnw
 /mvnw.cmd
+/.springBeans
