+++ news-service/src/main/java/com/news/article/NewsApplication.java
@@ -1,16 +1,14 @@
-package com.news;
+package com.news.article;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 
-@EnableEurekaClient
-@EnableZuulProxy
 @SpringBootApplication
-public class GatewayApplication {
+@EnableEurekaClient
+public class NewsApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GatewayApplication.class, args);
+		SpringApplication.run(NewsApplication.class, args);
 	}
 }
