+++ news-service/src/test/java/com/news/article/NewsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.news;
+package com.news.article;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class EurekaApplicationTests {
+public class NewsApplicationTests {
 
 	@Test
 	public void contextLoads() {
